Bump dependencies and upgrade Tauri
This commit updates dependencies in Cargo.toml and
frontend/package.json. It also upgrades Tauri and related dependencies
to the latest versions and updates pnpm-lock.yaml to reflect dependency
changes.

Updates include: - Upgrading `anyhow` to 1.0.99 - Upgrading `indexmap`
to 2.11.0 - Upgrading `bitflags` to 2.9.3 - Upgrading `objc2` to 0.6.2 -
Upgrading `brotli` to 8.0.2 - Upgrading `thiserror` to 2.0.16 -
Upgrading `clap` to 4.5.45 - Upgrading `cfg-if` to 1.0.3 - Upgrading
`cc` to 1.2.34 - Upgrading `form_urlencoded` to 1.2.2 - Upgrading `glob`
to 0.3.3 - Upgrading `hyper` to 1.7.0 - Upgrading `idna` to 1.1.0 -
Upgrading `io-uring` to 0.7.10 - Upgrading `libc` to 0.2.175 - Upgrading
`dlopen2` to 0.8.0 - Upgrading `percent-encoding` to 2.3.2 - Upgrading
`prettyplease` to 0.2.37 - Upgrading `proc-macro2` to 1.0.101 -
Upgrading `quick-xml` to 0.38.3 - Upgrading `regex` to 1.11.2 -
Upgrading `regex-automata` to 0.4.10 - Upgrading `regex-syntax` to 0.8.6
- Upgrading `reqwest` to 0.12.23 - Upgrading `serde-untagged` to 0.1.8 -
Upgrading `serde_json` to 1.0.143 - Upgrading `serialize-to-javascript`
to 0.1.2 - Upgrading `syn` to 2.0.106 - Upgrading `tao` to 0.34.2 -
Upgrading `tauri` to 2.8.3 - Upgrading `tauri-build` to 2.4.0 -
Upgrading `tauri-codegen` to 2.4.0 - Upgrading `tauri-macros` to 2.4.0 -
Upgrading `tauri-plugin` to 2.4.0 - Upgrading `tauri-plugin-opener` to
2.5.0 - Upgrading `tauri-runtime` to 2.8.0 - Upgrading
`tauri-runtime-wry` to 2.8.0 - Upgrading `tauri-utils` to 2.7.0 -
Upgrading `tempfile` to 3.21.0 - Upgrading `toml` to 0.9.5 - Upgrading
`winapi-util` to 0.1.10 - Upgrading `url` to 2.5.7 - Upgrading `uuid` to
1.18.0 - Upgrading `wry` to 0.53.1 - Upgrading `zbus` to 5.10.0 -
Upgrading `zvariant` to 5.7.0 - Upgrading `@tauri-apps/cli` to 2.8.2 -
Upgrading `@tauri-apps/plugin-dialog` to 2.3.3 - Upgrading
`@tauri-apps/plugin-opener` to 2.5.0 - Upgrading `@types/node` to 24.3.0
- Upgrading `vite` to 7.1.3